+++ src/Console/Application.php
@@ -518,4 +518,16 @@ class Application extends BaseApplication
 
         return true;
     }
+
+    public function extractNamespace(string $name, ?int $limit = null): string
+    {
+        $parts = explode(':', $name);
+
+        if ($limit === 1 && count($parts) >= 2 && $parts[0] === 'plugins') {
+            // Force grouping plugin commands
+            $limit = 2;
+        }
+
+        return implode(':', null === $limit ? $parts : \array_slice($parts, 0, $limit));
+    }
 }